backend/requirements.txt
# backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-multipart==0.0.18
torch>=2.0
torchvision>=0.15
timm>=1.0
grad-cam>=1.5
Pillow>=10.0
numpy>=1.24
